+++ src/main/java/org/stellar/sdk/TrustLineAsset.java
@@ -106,12 +106,12 @@ public abstract class TrustLineAsset implements Comparable<TrustLineAsset> {
       case ASSET_TYPE_CREDIT_ALPHANUM4:
         String assetCode4 =
             Util.paddedByteArrayToString(xdr.getAlphaNum4().getAssetCode().getAssetCode4());
-        accountId = StrKey.encodeStellarAccountId(xdr.getAlphaNum4().getIssuer());
+        accountId = StrKey.encodeEd25519PublicKey(xdr.getAlphaNum4().getIssuer());
         return TrustLineAsset.create(new AssetTypeCreditAlphaNum4(assetCode4, accountId));
       case ASSET_TYPE_CREDIT_ALPHANUM12:
         String assetCode12 =
             Util.paddedByteArrayToString(xdr.getAlphaNum12().getAssetCode().getAssetCode12());
-        accountId = StrKey.encodeStellarAccountId(xdr.getAlphaNum12().getIssuer());
+        accountId = StrKey.encodeEd25519PublicKey(xdr.getAlphaNum12().getIssuer());
         return TrustLineAsset.create(new AssetTypeCreditAlphaNum12(assetCode12, accountId));
       case ASSET_TYPE_POOL_SHARE:
         return new LiquidityPoolShareTrustLineAsset(